Consolidate the BeforeFinally and AfterFinally http operators
